# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/activate/license-type/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LicenseType"
                ],
                "summary": "添加许可证类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "添加许可证类型参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddLicenseType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加许可证类型",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/activate/license-type/del": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LicenseType"
                ],
                "summary": "删除许可证类型",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "类型ID",
                        "name": "type_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除许可证类型",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/activate/license-type/list": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LicenseType"
                ],
                "summary": "获取许可证类型列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "产品ID",
                        "name": "product_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码，从1开始",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取许可证类型列表",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
        "/activate/license-type/update-features": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LicenseType"
                ],
                "summary": "更新许可证类型功能列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "更新许可证类型功能列表参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateLicenseTypeFeatures"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新许可证类型功能列表",
                        "schema": {
                            "$ref": "#/definitions/resp.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AddLicenseType": {
            "type": "object",
            "required": [
                "license_type",
                "product_id",
                "type_name"
            ],
            "properties": {
                "feature_ids": {
                    "description": "功能ID列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "license_type": {
                    "description": "许可证编码",
                    "type": "string"
                },
                "product_id": {
                    "description": "产品ID",
                    "type": "integer"
                },
                "type_name": {
                    "description": "许可证类型名称",
                    "type": "string"
                }
            }
        },
...
        "dto.UpdateLicenseTypeFeatures": {
            "type": "object",
            "required": [
                "type_id"
            ],
            "properties": {
                "feature_ids": {
                    "description": "功能ID列表",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "type_id": {
                    "description": "许可证类型ID",
                    "type": "integer"
                }
            }
        },
...
        "resp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    }
}